build.rs: correct dynamic library name for Linux release build (#5)
* build.rs: correct dynamic library name for Linux release build

Based on the context and the consistent usage of libprojectM-4 in the
rest of the file, it appears that the equal sign (=) is a typo and
should be replaced with a hyphen-minus (-). Current naming of the
dynamic lib causes link failure when building on Linux with release
profile.

* build.rs: remove lib prefix from dylib name on Linux release profile

On Linux systems, the convention is to use the library name without the
lib prefix. This change ensures that the correct library name is used
during the build process on Linux platforms.